requirements.txt
sqlalchemy==1.4.42
python-dotenv==0.21.0